Use a non-sampling method for determining accuracy
Behdad suggested this new method a while ago, which is faster and more
accurate than sampling. The biggest issue though is that it's harder
to report a good error for the approximation. I tried to deal with
this by keeping the curve-to-spline distance function, which a user
can run to get the error which would have been returned before.